// src/features/division-builder/calculation-types.ts
export type UnitStats = {
  organization: number;
  hp: number;
  recoveryRate: number;
  suppression: number;
  weight: number;
  supplyUse: number;
  fuelUse: number;
  speed: number;
  softAttack: number;
  hardAttack: number;
  airAttack: number;
  defense: number;
  breakthrough: number;
  armor: number;
  piercing: number;
  hardness: number;
  entrenchment: number;
};